Extra scene work.
The scene now contains 4 buttons, of which one is blinking.
For this to work I had to change the timer in Spaceship from a long to either a float or a double (i chose float, because of repetetive addition), because long lacks the precision to keep track of the updates and the timer stays 0.

+++ src/main/java/nl/tudelft/pixelperfect/route/Route.java
@@ -51,7 +51,7 @@ public class Route {
    *          Current timestamp in milliseconds.
    * @return Boolean.
    */
-  public boolean isCompleted(long now) {
+  public boolean isCompleted(float now) {
     return (now >= timestamp + duration);
   }
